git-audit-related: extract hashes of related commits
For a given list of revs, extract all

	Related: <hash>

labels in commit log messages. Extract the hashes and resolve
them to their full form. Also, if an extracted hash is not
reachable between the given start rev and the current rev
in the list, try to resolve it via reflog (e.g. after some
rebase) thus the one included between start and current rev
is used instead of the obsolete one.

Example (2 related hashes extracted and resolved):

commit abcdef0123456789abcdef0123456789abcdef01
abcabcabcabc0101010156565656567878787878
5678901234fabcde5678901234fabcde567890ab

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>