git-audit-log: show audit commits in console
The git-audit-log wraps git log command to present the audit
logs in a better way to the user. It hides the diff of audit
commits and replaces them with a code snipped highlighting the
audited changes.

An audit commit is recognized from the diff chunk information.
If the source offset and length is same as destination offset
and length of a chunk, such chunk is considered as an audit
commit diff. Other chunks are treated as suggestion commit diffs.

The command outputs colors and uses its own pager (less -R).
It can be configured via environment variables:

  GIT_AUDIT_COLOR - 0 to disable colors
  GIT_AUDIT_PAGER - command to use as a pager (e.g. "cat")

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>