git-audit-report: generate HTML audit reports
The command produces (quite) a huge XML file that is post-processed
by a XSLT processor. It assumes xsltproc as default. The provided
git-audit-report.xsl is a transformation that generates a HTML
report from the collected information about a range of audit commits.

Example usage:

  $ git checkout -b audit-master-branch origin/master
  ...
  $ git commit
  ...
  $ git commit
  ...
  $ git-audit-report origin/master

Result can be found in report.html. It is highly recommended to call
the command from the root of the git repository (no checking is
implemented yet). Also, the report.xml file is created.

The HTML report is JavaScript-free and uses only some CSS to highlight
the audited code chunks (compare to git-audit-log). Also, git blame
data are included and available from the HTML page. The report.html
is standalone.

There is one environment variable available for debugging:

  GIT_AUDIT_REPORT_DEBUG:
  	when 1, output debug logs
	when 2, use `dirname $0` as basename of scripts and binaries

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>